format zero padded durations like the dashboard
