+++ packages/desktop-gui/package.json
@@ -44,7 +44,6 @@
     "react-bootstrap-modal": "3.0.1",
     "react-dom": "^15.6.1",
     "react-loader": "^2.4.0",
-    "string.prototype.padstart": "^3.0.0",
     "zunder": "5.6.5"
   }
 }
